requirements.txt
streamlit
gspread
oauth2client
pandas
rapidfuzz>=3.0.0,<4.0.0
inflect
python-dotenv
together
recipe-scrapers